all prototypes checked. all funtions work with francinette (--strict)

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: mdoll <mdoll@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/19 15:08:48 by mdoll             #+#    #+#             */
-/*   Updated: 2022/12/19 15:23:13 by mdoll            ###   ########.fr       */
+/*   Updated: 2022/12/22 14:06:19 by mdoll            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (s[i])
